[MRG] Set dtype integer

+++ skopt/utils.py
@@ -490,13 +490,15 @@ def normalize_dimensions(dimensions):
                 transformed_dimensions.append(
                     Real(dimension.low, dimension.high, dimension.prior,
                          name=dimension.name,
-                         transform="normalize")
+                         transform="normalize",
+                         dtype=dimension.dtype)
                     )
             elif isinstance(dimension, Integer):
                 transformed_dimensions.append(
                     Integer(dimension.low, dimension.high,
                             name=dimension.name,
-                            transform="normalize")
+                            transform="normalize",
+                            dtype=dimension.dtype)
                     )
             else:
                 raise RuntimeError("Unknown dimension type "
